Move utils in preparation for ritm migration
