+++ web_ui/packages/smart-tools/package.json
@@ -6,7 +6,8 @@
   "main": "index.ts",
   "type": "module",
   "dependencies": {
-    "comlink": "^4.3.1"
+    "comlink": "^4.3.1",
+    "@doodle3d/clipper-js": "^1.0.11"
   },
   "devDependencies": {
     "@geti/config": "*",
