+++ web_ui/src/webworkers/ritm.worker.ts
@@ -3,7 +3,15 @@
 
 // Dependencies get bundled into the worker
 
-import { approximateShape, OpenCVLoader } from '@geti/smart-tools';
+import {
+    approximateShape,
+    concatFloat32Arrays,
+    isPolygonValid,
+    loadSource,
+    OpenCVLoader,
+    stackPlanes,
+} from '@geti/smart-tools';
+import { Point, Polygon, Shape, ShapeType } from '@geti/smart-tools/src/shared/interfaces';
 import { expose } from 'comlink';
 import ndarray from 'ndarray';
 import ops from 'ndarray-ops';
@@ -11,17 +19,8 @@ import * as ort from 'onnxruntime-web';
 import type OpenCVTypes from 'OpenCVTypes';
 
 import { RegionOfInterest } from '../core/annotations/annotation.interface';
-import { Point, Shape } from '../core/annotations/shapes.interface';
-import { ShapeType } from '../core/annotations/shapetype.enum';
-import {
-    RITMContour,
-    RITMMethods,
-    RITMPoint,
-    TEMPLATE_SIZE,
-} from '../pages/annotator/tools/ritm-tool/ritm-tool.interface';
-import { isPolygonValid } from '../pages/annotator/tools/utils';
+import { RITMContour, RITMPoint, TEMPLATE_SIZE } from '../pages/annotator/tools/ritm-tool/ritm-tool.interface';
 import { sessionParams } from '../pages/annotator/tools/wasm-utils';
-import { concatFloat32Arrays, loadSource, stackPlanes } from './utils';
 
 declare const self: DedicatedWorkerGlobalScope;
 
@@ -42,7 +41,7 @@ interface Models {
     main: ort.InferenceSession;
 }
 
-class RITM implements RITMMethods {
+class RITM {
     models: Models | undefined;
     image: OpenCVTypes.Mat | undefined;
     mask: OpenCVTypes.Mat | undefined;
@@ -141,18 +140,18 @@ class RITM implements RITMMethods {
         }
 
         switch (outputShape) {
-            case ShapeType.Polygon:
-                const shape: Shape = { shapeType: ShapeType.Polygon, points: contour.contour };
+            case 'polygon':
+                const shape: Polygon = { shapeType: 'polygon', points: contour.contour };
                 if (isPolygonValid(shape)) {
                     return shape;
                 }
                 return undefined;
-            case ShapeType.RotatedRect:
+            case 'rotated-rect':
                 const { x, y } = contour.minAreaRect.center;
                 const { width, height } = contour.minAreaRect.size;
                 const angle = contour.minAreaRect.angle;
 
-                return { shapeType: ShapeType.RotatedRect, x, y, width, height, angle };
+                return { shapeType: 'rotated-rect', x, y, width, height, angle };
         }
         throw 'Not implemented shape.';
     }
